Enums

Create enum sets

Usage

Enum(name, ...)

enum(values, ...)

Arguments

name

The name of the `enum`

values, ...

Set values for the `enum`. These can be names or unnamed.

Value

An `enum` object, which is a function with S3 methods

Details

`enum`s are sets of named or unnamed values. The `enum` object has built-in features to retrieve values by name, provide a list of all values,

Examples

Letters <- enum(letters)
Letters()        # when missing, returns list of everything
#> $a
#> [1] "a"
#> 
#> $b
#> [1] "b"
#> 
#> $c
#> [1] "c"
#> 
#> $d
#> [1] "d"
#> 
#> $e
#> [1] "e"
#> 
#> $f
#> [1] "f"
#> 
#> $g
#> [1] "g"
#> 
#> $h
#> [1] "h"
#> 
#> $i
#> [1] "i"
#> 
#> $j
#> [1] "j"
#> 
#> $k
#> [1] "k"
#> 
#> $l
#> [1] "l"
#> 
#> $m
#> [1] "m"
#> 
#> $n
#> [1] "n"
#> 
#> $o
#> [1] "o"
#> 
#> $p
#> [1] "p"
#> 
#> $q
#> [1] "q"
#> 
#> $r
#> [1] "r"
#> 
#> $s
#> [1] "s"
#> 
#> $t
#> [1] "t"
#> 
#> $u
#> [1] "u"
#> 
#> $v
#> [1] "v"
#> 
#> $w
#> [1] "w"
#> 
#> $x
#> [1] "x"
#> 
#> $y
#> [1] "y"
#> 
#> $z
#> [1] "z"
#> 
Letters(Letters) # pass the enum itself to retrieve just the values
#>  [1] "a" "b" "c" "d" "e" "f" "g" "h" "i" "j" "k" "l" "m" "n" "o" "p" "q" "r" "s"
#> [20] "t" "u" "v" "w" "x" "y" "z"

# value, ... are combined:
Continents <- enum(
  "Africa",
  "Antarctica",
  "Asia",
  "Europe",
  "North America",
  "Oceania",
  "South America"
)

Continents()
#> $Africa
#> [1] "Africa"
#> 
#> $Antarctica
#> [1] "Antarctica"
#> 
#> $Asia
#> [1] "Asia"
#> 
#> $Europe
#> [1] "Europe"
#> 
#> $`North America`
#> [1] "North America"
#> 
#> $Oceania
#> [1] "Oceania"
#> 
#> $`South America`
#> [1] "South America"
#> 
Continents@Asia
#> [1] "Asia"
Continents("Asia")
#> [1] "Asia"
Continents(c("Asia", "Europe"))
#> [1] "Asia"   "Europe"
try(Continents("Atlantis"))
#> Error : 'Atlantis' is not a valid enum value

Colors := Enum(c(
  Red = "#FF0000",
  Green = "#00FF00",
  Blue = "#0000FF",
  Yellow = "#FFFF00",
  Black = "#000000",
  White = "#FFFFFF",
  Purple = "#800080",
  Orange = "#FFA500",
  Pink = "#FFC0CB",
  Brown = "#A52A2A",
  Gray = "#808080",
  Cyan = "#00FFFF",
  Magenta = "#FF00FF"
))

Colors@Red
#> [1] "#FF0000"
Colors(c("Red", "Blue"))
#> [1] "#FF0000" "#0000FF"
Colors("Green")
#> [1] "#00FF00"
try(Colors("Turquoise"))  # This will raise an error
#> Error : 'Turquoise' is not a valid Enum{Colors} value

Shapes <- enum(c(
  Circle = "A round shape",
  Square = "A shape with four equal sides",
  Triangle = "A shape with three sides",
  Rectangle = "A shape with four sides and equal opposite sides",
  Oval = "An elongated circle",
  Star = "A shape with points radiating from a center",
  Heart = "A shape symbolizing love"
))

Shapes@Circle
#> [1] "A round shape"
Shapes(c("Circle", "Square"))
#> [1] "A round shape"                 "A shape with four equal sides"
Colors(Colors)
#>  [1] "#FF0000" "#00FF00" "#0000FF" "#FFFF00" "#000000" "#FFFFFF" "#800080"
#>  [8] "#FFA500" "#FFC0CB" "#A52A2A" "#808080" "#00FFFF" "#FF00FF"
print((function(x = Colors) Colors(x))())
#>  [1] "#FF0000" "#00FF00" "#0000FF" "#FFFF00" "#000000" "#FFFFFF" "#800080"
#>  [8] "#FFA500" "#FFC0CB" "#A52A2A" "#808080" "#00FFFF" "#FF00FF"
print((function(x = Colors) Colors(x))("Red"))
#> [1] "#FF0000"